refactor(home): extract landing page sections into composable components (#677)
