+++ src/components/landing/SignOutDialog.tsx
@@ -1,0 +1,48 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+
+type SignOutDialogProps = {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  onSignOut: () => void | Promise<void>;
+};
+
+export default function SignOutDialog({
+  open,
+  onOpenChange,
+  onSignOut,
+}: SignOutDialogProps) {
+  return (
+    <AlertDialog open={open} onOpenChange={onOpenChange}>
+      <AlertDialogContent className="bg-white dark:bg-slate-900 border-slate-200 dark:border-slate-800 text-slate-900 dark:text-white rounded-2xl animate-in fade-in-50 zoom-in-95 duration-200">
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you sure you want to sign out?</AlertDialogTitle>
+          <AlertDialogDescription className="text-slate-500 dark:text-slate-400">
+            You will need to log in again to access your classroom insights and
+            doubt-solving history.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel className="bg-slate-100 dark:bg-slate-800 border-slate-200 dark:border-slate-700 text-slate-700 dark:text-slate-300 hover:bg-slate-200 dark:hover:bg-slate-700 rounded-xl transition-all duration-200">
+            Cancel
+          </AlertDialogCancel>
+          <AlertDialogAction
+            onClick={onSignOut}
+            className="bg-red-600 hover:bg-red-700 text-white border-none rounded-xl transition-all duration-200"
+          >
+            Sign Out
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
+
